review: move PolledReader to api/ boundary; align 044 spec to i64
Resolves two CodeRabbit findings on PR #51.

1. (Major) `src/api/*` was importing `PolledReader` and `Unsubscribe`
   from `src/hooks/usePolledReader`, inverting the CLAUDE.md dependency
   direction (`hooks/` calls `api/`, not the other way). The contract
   the api/ readers implement belongs in api/.

   - New `src/api/polledReader.ts` owns `PolledReader<T>` + `Unsubscribe`
   - `src/hooks/usePolledReader.ts` keeps `usePolledReader<T>` hook and
     `PolledState<T>` discriminated union; imports `PolledReader` from
     api/polledReader
   - All 6 api/ files now import from `./polledReader`; no api/ file
     references hooks/
   - Components and the hook tests import the type from `api/polledReader`
     and the hook from `hooks/usePolledReader`

2. (Minor) Issue 044's spec block showed `now_secs: u64`; implementation
   uses `i64`. Unix timestamps are signed; matched the spec to the impl
   (no code change).

`task check` green. 69/69 frontend tests pass.
