+++ src/api/polledReader.ts
@@ -1,0 +1,6 @@
+export type Unsubscribe = () => void;
+
+export interface PolledReader<T> {
+  read(): Promise<T>;
+  subscribe?(onChange: () => void): Unsubscribe | Promise<Unsubscribe>;
+}
